pyproject.toml
[tool.poetry]
name = "gitlab2jiraissues"
version = "0.1.0"
description = ""
authors = ["Rasmus Paltschik"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
invoke = "^1.7.3"
typing-extensions = "^4.4.0"
requests = "^2.28.1"
python-dotenv = "^0.21.0"
pandas = "^1.5.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
coverage = "^6.5.0"
pylint = "^2.15.5"
autopep8 = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
